feat: implement CompleteQuiz API endpoint

+++ quiz-server/Modules/Quiz/Modules.Quiz/Domain/QuizAttempt.cs
@@ -8,7 +8,7 @@ namespace Modules.Quiz.Domain
         public Guid UserId { get; set; }
         public DateTime StartedAt { get; set; }
         public DateTime? CompletedAt { get; set; }
-        public int? Score { get; set; }
+        public double? Score { get; set; }
         public bool? IsPassed { get; set; }
         public TimeSpan? TimeSpent { get; set; }
         public string Status { get; set; } = "in-progress"; // in-progress, completed, abandoned
